[pc-12883][api] moves function to subscription api
