+++ api/src/pcapi/core/subscription/api.py
@@ -435,7 +435,7 @@ def get_maintenance_page_type(user: users_models.User) -> typing.Optional[models
 
 
 def activate_beneficiary_if_no_missing_step(user: users_models.User) -> None:
-    if users_api.has_passed_all_checks_to_become_beneficiary(user):
+    if has_passed_all_checks_to_become_beneficiary(user):
         activate_beneficiary(user)  # calls update_external_user
     else:
         users_external.update_external_user(user)
@@ -523,3 +523,24 @@ def update_user_birth_date(user: users_models.User, birth_date: typing.Optional[
     if user.dateOfBirth != birth_date and birth_date is not None:
         user.dateOfBirth = birth_date
         pcapi_repository.repository.save(user)
+
+
+def has_passed_all_checks_to_become_beneficiary(user: users_models.User) -> bool:
+    fraud_check = users_api.get_activable_identity_fraud_check(user)
+    if not fraud_check:
+        return False
+
+    if (
+        not user.is_phone_validated
+        and fraud_check.eligibilityType != users_models.EligibilityType.UNDERAGE
+        and FeatureToggle.FORCE_PHONE_VALIDATION.is_active()
+    ):
+        return False
+
+    if not fraud_api.has_performed_honor_statement(user, fraud_check.eligibilityType):
+        if not FeatureToggle.IS_HONOR_STATEMENT_MANDATORY_TO_ACTIVATE_BENEFICIARY.is_active():
+            logger.warning("The honor statement has not been performed or recorded", extra={"user_id": user.id})
+        else:
+            return False
+
+    return True
